updated project with jwt-extended

+++ resources/item.py
@@ -1,6 +1,11 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
-
+from flask_jwt_extended import (
+    jwt_required, 
+    get_jwt_claims, 
+    jwt_optional, 
+    get_jwt_identity, 
+    fresh_jwt_required
+)
 from models.item import ItemModel
 
 
@@ -26,6 +31,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required # The fresh flag is set to True (as done by the /login end-point)
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "Item already exists"}, 400
@@ -40,7 +46,11 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return{'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -62,5 +72,12 @@ class Item(Resource):
 
 
 class Items(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items =  [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+
+        return {'items': [item['name'] for item in items],
+                'message': 'More data available if you log in.'}, 200
